fix(nodejs-package): skip native frames less agressively
Previous algorithm would skip any native or kernel frame between two JS
frames. This is very agressive and would lead to unwanted skipped
frames. By changing to a list of frames which are safe to skip, we can
generate more relevant Node.js Package FlameGraphs. We can grow this
list as we learn more frames which can be included.